+++ src/rules/no-restricted-paths.js
@@ -29,8 +29,28 @@ module.exports = {
             items: {
               type: 'object',
               properties: {
-                target: { type: 'string' },
-                from: { type: 'string' },
+                target: {
+                  oneOf: [
+                    { type: 'string' },
+                    {
+                      type: 'array',
+                      items: { type: 'string' },
+                      uniqueItems: true,
+                      minLength: 1,
+                    },
+                  ],
+                },
+                from: {
+                  oneOf: [
+                    { type: 'string' },
+                    {
+                      type: 'array',
+                      items: { type: 'string' },
+                      uniqueItems: true,
+                      minLength: 1,
+                    },
+                  ],
+                },
                 except: {
                   type: 'array',
                   items: {
@@ -56,19 +76,27 @@ module.exports = {
     const basePath = options.basePath || process.cwd();
     const currentFilename = context.getPhysicalFilename ? context.getPhysicalFilename() : context.getFilename();
     const matchingZones = restrictedPaths.filter((zone) => {
-      const targetPath = path.resolve(basePath, zone.target);
+      return [].concat(zone.target)
+        .map(target => path.resolve(basePath, target))
+        .some(targetPath => isMatchingTargetPath(currentFilename, targetPath));
+    });
 
+    function isMatchingTargetPath(filename, targetPath) {
       if (isGlob(targetPath)) {
-        return minimatch(currentFilename, targetPath);
+        return minimatch(filename, targetPath);
       }
 
-      return containsPath(currentFilename, targetPath);
-    });
+      return containsPath(filename, targetPath);
+    }
 
     function isValidExceptionPath(absoluteFromPath, absoluteExceptionPath) {
       const relativeExceptionPath = path.relative(absoluteFromPath, absoluteExceptionPath);
 
       return importType(relativeExceptionPath, context) !== 'parent';
+    }
+
+    function areBothGlobPatternAndAbsolutePath(areGlobPatterns) {
+      return areGlobPatterns.some((isGlob) => isGlob) && areGlobPatterns.some((isGlob) => !isGlob);
     }
 
     function reportInvalidExceptionPath(node) {
@@ -78,49 +106,41 @@ module.exports = {
       });
     }
 
-    function reportInvalidExceptionGlob(node) {
+    function reportInvalidExceptionMixedGlobAndNonGlob(node) {
       context.report({
         node,
-        message: 'Restricted path exceptions must be glob patterns when `from` is a glob pattern',
+        message: 'Restricted path `from` must contain either only glob patterns or none',
       });
     }
 
-    const makePathValidator = (zoneFrom, zoneExcept = []) => {
-      const absoluteFrom = path.resolve(basePath, zoneFrom);
-      const isGlobPattern = isGlob(zoneFrom);
-      let isPathRestricted;
-      let hasValidExceptions;
+    function reportInvalidExceptionGlob(node) {
+      context.report({
+        node,
+        message: 'Restricted path exceptions must be glob patterns when `from` contains glob patterns',
+      });
+    }
+
+    function computeMixedGlobAndAbsolutePathValidator() {
+      return {
+        isPathRestricted: () => true,
+        hasValidExceptions: false,
+        reportInvalidException: reportInvalidExceptionMixedGlobAndNonGlob,
+      };
+    }
+
+    function computeGlobPatternPathValidator(absoluteFrom, zoneExcept) {
       let isPathException;
-      let reportInvalidException;
 
-      if (isGlobPattern) {
-        const mm = new Minimatch(absoluteFrom);
-        isPathRestricted = (absoluteImportPath) => mm.match(absoluteImportPath);
+      const mm = new Minimatch(absoluteFrom);
+      const isPathRestricted = (absoluteImportPath) => mm.match(absoluteImportPath);
+      const hasValidExceptions = zoneExcept.every(isGlob);
 
-        hasValidExceptions = zoneExcept.every(isGlob);
-
-        if (hasValidExceptions) {
-          const exceptionsMm = zoneExcept.map((except) => new Minimatch(except));
-          isPathException = (absoluteImportPath) => exceptionsMm.some((mm) => mm.match(absoluteImportPath));
-        }
-
-        reportInvalidException = reportInvalidExceptionGlob;
-      } else {
-        isPathRestricted = (absoluteImportPath) => containsPath(absoluteImportPath, absoluteFrom);
-
-        const absoluteExceptionPaths = zoneExcept
-          .map((exceptionPath) => path.resolve(absoluteFrom, exceptionPath));
-        hasValidExceptions = absoluteExceptionPaths
-          .every((absoluteExceptionPath) => isValidExceptionPath(absoluteFrom, absoluteExceptionPath));
-
-        if (hasValidExceptions) {
-          isPathException = (absoluteImportPath) => absoluteExceptionPaths.some(
-            (absoluteExceptionPath) => containsPath(absoluteImportPath, absoluteExceptionPath),
-          );
-        }
-
-        reportInvalidException = reportInvalidExceptionPath;
+      if (hasValidExceptions) {
+        const exceptionsMm = zoneExcept.map((except) => new Minimatch(except));
+        isPathException = (absoluteImportPath) => exceptionsMm.some((mm) => mm.match(absoluteImportPath));
       }
+
+      const reportInvalidException = reportInvalidExceptionGlob;
 
       return {
         isPathRestricted,
@@ -128,6 +148,66 @@ module.exports = {
         isPathException,
         reportInvalidException,
       };
+    }
+
+    function computeAbsolutePathValidator(absoluteFrom, zoneExcept) {
+      let isPathException;
+
+      const isPathRestricted = (absoluteImportPath) => containsPath(absoluteImportPath, absoluteFrom);
+
+      const absoluteExceptionPaths = zoneExcept
+        .map((exceptionPath) => path.resolve(absoluteFrom, exceptionPath));
+      const hasValidExceptions = absoluteExceptionPaths
+        .every((absoluteExceptionPath) => isValidExceptionPath(absoluteFrom, absoluteExceptionPath));
+
+      if (hasValidExceptions) {
+        isPathException = (absoluteImportPath) => absoluteExceptionPaths.some(
+          (absoluteExceptionPath) => containsPath(absoluteImportPath, absoluteExceptionPath),
+        );
+      }
+
+      const reportInvalidException = reportInvalidExceptionPath;
+
+      return {
+        isPathRestricted,
+        hasValidExceptions,
+        isPathException,
+        reportInvalidException,
+      };
+    }
+
+    function reportInvalidExceptions(validators, node) {
+      validators.forEach(validator => validator.reportInvalidException(node));
+    }
+
+    function reportImportsInRestrictedZone(validators, node, importPath, customMessage) {
+      validators.forEach(() => {
+        context.report({
+          node,
+          message: `Unexpected path "{{importPath}}" imported in restricted zone.${customMessage ? ` ${customMessage}` : ''}`,
+          data: { importPath },
+        });
+      });
+    }
+
+    const makePathValidators = (zoneFrom, zoneExcept = []) => {
+      const allZoneFrom = [].concat(zoneFrom);
+      const areGlobPatterns = allZoneFrom.map(isGlob);
+
+      if (areBothGlobPatternAndAbsolutePath(areGlobPatterns)) {
+        return [computeMixedGlobAndAbsolutePathValidator()];
+      }
+
+      const isGlobPattern = areGlobPatterns.every((isGlob) => isGlob);
+
+      return allZoneFrom.map(singleZoneFrom => {
+        const absoluteFrom = path.resolve(basePath, singleZoneFrom);
+
+        if (isGlobPattern) {
+          return computeGlobPatternPathValidator(absoluteFrom, zoneExcept);
+        }
+        return computeAbsolutePathValidator(absoluteFrom, zoneExcept);
+      });
     };
 
     const validators = [];
@@ -141,35 +221,18 @@ module.exports = {
 
       matchingZones.forEach((zone, index) => {
         if (!validators[index]) {
-          validators[index] = makePathValidator(zone.from, zone.except);
+          validators[index] = makePathValidators(zone.from, zone.except);
         }
 
-        const {
-          isPathRestricted,
-          hasValidExceptions,
-          isPathException,
-          reportInvalidException,
-        } = validators[index];
+        const applicableValidatorsForImportPath = validators[index].filter(validator => validator.isPathRestricted(absoluteImportPath));
 
-        if (!isPathRestricted(absoluteImportPath)) {
-          return;
-        }
+        const validatorsWithInvalidExceptions = applicableValidatorsForImportPath.filter(validator => !validator.hasValidExceptions);
+        reportInvalidExceptions(validatorsWithInvalidExceptions, node);
 
-        if (!hasValidExceptions) {
-          reportInvalidException(node);
-          return;
-        }
-
-        const pathIsExcepted = isPathException(absoluteImportPath);
-        if (pathIsExcepted) {
-          return;
-        }
-
-        context.report({
-          node,
-          message: `Unexpected path "{{importPath}}" imported in restricted zone.${zone.message ? ` ${zone.message}` : ''}`,
-          data: { importPath },
-        });
+        const applicableValidatorsForImportPathExcludingExceptions = applicableValidatorsForImportPath
+          .filter(validator => validator.hasValidExceptions)
+          .filter(validator => !validator.isPathException(absoluteImportPath));
+        reportImportsInRestrictedZone(applicableValidatorsForImportPathExcludingExceptions, node, importPath, zone.message);
       });
     }